+++ src/GA/ga.py
@@ -1,13 +1,6 @@
-import datetime
-import copy
-import pandas as pd
-import random
-
-from src.chromosome import *
-from src.pmsbx_ga import _generate_parent
-from src.utils import *
-from src.conditions_ga import cal_fitness_value
-from src.draw_chart import draw_fitness
+from src.PMSBX_GA.pmsbx_ga import _generate_parent
+from src.utilities.utils import *
+from src.GA.conditions_ga import cal_fitness_value
 
 """Create population function - GA"""
