+++ src/NSGA/nsga.py
@@ -1,13 +1,8 @@
-import datetime
 import copy
-import pandas as pd
-import random
 
-from src.chromosome import *
-from src.pmsbx_ga import _generate_parent
-from src.utils import *
-from src.conditions_ga import cal_fitness_value
-from src.utils import _cal_end_date
+from src.PMSBX_GA.pmsbx_ga import _generate_parent
+from src.utilities.utils import *
+from src.utilities.utils import _cal_end_date
 
 # # bit_dict_team
 # # ======================MODIFY======================
